refactored main.js into module
