+++ extension/secureflow/src/ui/webview/index.html
@@ -142,6 +142,6 @@
             <div id="profileContent"></div>
         </div>
     </div>
-    <script src="${scriptUri}"></script>
+    <script type="module" src="${scriptUri}"></script>
 </body>
 </html>
